--- src/main/groovy/com/deploygate/gradle/plugins/DeployGateAllUploadTask.groovy
@@ -1,11 +1,0 @@
-package com.deploygate.gradle.plugins
-
-import org.gradle.api.tasks.TaskAction
-
-class DeployGateAllUploadTask extends DeployGateTask {
-    @TaskAction
-    def uploadDeployGate() {
-        List<Apk> apks = Apk.getApks(project)
-        super.upload(project, apks)
-    }
-}
