+++ src/main/groovy/com/deploygate/gradle/plugins/tasks/DeployGateSetupCredentialTask.groovy
@@ -1,12 +1,15 @@
-package com.deploygate.gradle.plugins
+package com.deploygate.gradle.plugins.tasks
 
+import com.deploygate.gradle.plugins.Config
+import com.deploygate.gradle.plugins.auth.DeployGateLocalCredential
+import com.deploygate.gradle.plugins.utils.UrlUtils
 import com.sun.net.httpserver.HttpExchange
 import com.sun.net.httpserver.HttpHandler
 import com.sun.net.httpserver.HttpServer
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
 
-import java.awt.Desktop
+import java.awt.*
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
