+++ src/test/groovy/com/deploygate/gradle/plugins/ApplyPluginTest.groovy
@@ -1,5 +1,6 @@
 package com.deploygate.gradle.plugins
 
+import com.deploygate.gradle.plugins.extensions.DeployGateExtension
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
 import org.junit.Test
